refactor: organiza arquivos e pastas por módulos e services
