+++ popup.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Taiga Sprint Burndown v.0.3</title>
-    <script type="module" src="src/popup/index.ts"></script>
+    <script type="module" src="src/modules/popup/index.ts"></script>
     <link rel="stylesheet" href="styles.css">
   </head>
   <body>
