--- src/content/task/getTasks.ts
@@ -1,17 +1,0 @@
-import { Task } from "../../interfaces";
-import { getTaskInfo } from "./getTaskInfo";
-
-export function getTasks(document: Element): Task[] {
-  try {
-    const tasks: Task[] = [];
-    document.querySelectorAll('.card-inner')
-      .forEach((item) => {
-        tasks.push(getTaskInfo(item))
-      })
-
-    
-    return tasks;
-  } catch (error) {
-    return [];
-  }
-}
