+++ src/interfaces/Member.ts
@@ -1,0 +1,9 @@
+export interface MemberTaskInfo {
+  member: string;
+  img: string;
+  assignedTasks: number;
+  closedTasks: number;
+  assignedHours: string;
+  closedHours: string;
+  hoursPerDay: string;
+}
