+++ src/modules/background/index.ts
