+++ src/modules/content/helpers/duration/getDuration.ts
@@ -1,0 +1,10 @@
+export function getDuration() {
+  try {
+    const duration = (
+      document.querySelector(".taskboard-header h1 .date") as HTMLElement
+    ).innerText;
+    return duration;
+  } catch (error) {
+    return "NOT FOUND";
+  }
+}
