+++ src/modules/content/helpers/percentage/calculatePercentage.ts
