+++ src/modules/content/layout/duration.ts
