+++ src/modules/content/layout/qtdNew.ts
