+++ src/modules/content/layout/totalClosed.ts
