+++ src/modules/content/services/squadService.ts
@@ -1,0 +1,16 @@
+export const squadService = {
+  getSquadName(): string {
+    try {
+      const squadName = (
+        document.querySelector(".taskboard-header h1 span") as HTMLElement
+      ).innerText
+        .split("-")[0]
+        .trim()
+        .substring(3);
+
+      return squadName;
+    } catch (error) {
+      return "NOT FOUND";
+    }
+  },
+};
